refactor: unify file naming and placement conventions
- Rename one-time transaction action.ts files to actions.ts (CLAUDE.md
  convention)
- Move page-specific Dashboard component from shared components/ into
  app/(protected)/dashboard/_components/
- Flatten transactions/recurring/types/index.ts to types.ts
